Refactor: 하나의 Console, 하나의 Controller로 통합한다.
