+++ src/main/java/org/prgms/springbootbasic/controller/MainController.java
@@ -2,18 +2,22 @@ package org.prgms.springbootbasic.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.prgms.springbootbasic.common.console.Console;
-import org.prgms.springbootbasic.common.console.MainConsole;
 import org.prgms.springbootbasic.domain.VoucherType;
 import org.prgms.springbootbasic.domain.customer.Customer;
 import org.prgms.springbootbasic.domain.voucher.VoucherPolicy;
 import org.prgms.springbootbasic.service.CustomerService;
 import org.prgms.springbootbasic.service.VoucherService;
+import org.prgms.springbootbasic.service.WalletService;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Controller;
 
 import java.util.InputMismatchException;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.UUID;
+
+import static org.prgms.springbootbasic.common.console.Console.*;
+import static org.prgms.springbootbasic.common.console.Console.typeVoucherId;
 
 @Controller
 @Slf4j
@@ -23,26 +27,31 @@ public class MainController {
     private static final String CREATE = "create";
     private static final String BLACK = "black";
     private static final String WALLET = "wallet";
+    private static final String ALLOCATE = "allocate";
+    private static final String DELETE = "delete";
+    private static final String SHOW_CUSTOMER_BY_VOUCHER = "showVoucherByCustomer";
+    private static final String SHOW_VOUCHER_BY_CUSTOMER = "showCustomerByVoucher";
+    private static final String BACK = "back";
 
     private final VoucherService voucherService;
     private final CustomerService customerService;
-    private final WalletController walletController;
+    private final WalletService walletService;
 
-    public MainController(VoucherService voucherService, CustomerService customerService, WalletController walletController) {
+    public MainController(VoucherService voucherService, CustomerService customerService, WalletService walletService) {
         this.voucherService = voucherService;
         this.customerService = customerService;
-        this.walletController = walletController;
+        this.walletService = walletService;
     }
 
     public void run() {
             String command = "";
             while (!command.equals(EXIT)) {
                 try {
-                    command = MainConsole.readCommand();
+                    command = Console.readCommand();
 
                     executeCommand(command);
                 } catch (InputMismatchException e) {
-                    String invalidVal = MainConsole.ignoreLine();
+                    String invalidVal = Console.ignoreLine();
 
                     log.warn("User input = {}", invalidVal);
                     throw new IllegalArgumentException("Not integer.");
@@ -53,22 +62,22 @@ public class MainController {
                     log.error("Scanner is closed");
                     throw new RuntimeException("Scanner is closed.");
                 } catch (IllegalArgumentException e) {
-                    MainConsole.printArgException();
+                    Console.printArgException();
                 } catch (DataAccessException e) {
                     log.error("Database error.", e);
-                    MainConsole.printRuntimeException();
+                    Console.printRuntimeException();
                 } catch (RuntimeException e) {
-                    MainConsole.printRuntimeException();
+                    Console.printRuntimeException();
                 }
             }
-    }
+    } // 프론트 컨트롤러 패턴에 대해. 공부해보자. 컨트롤러 일관성 없음. 왜 wallet만 따로 존재하는? 고민... -> 그냥 통일합시다.
 
     private void executeCommand(String command) {
         switch (command) {
             case CREATE -> create();
             case LIST -> list();
             case BLACK -> black();
-            case WALLET -> wallet();
+            case WALLET -> runWallet();
             case EXIT -> {}
             default -> {
                 log.warn("invalid command. now command = {}", command);
@@ -78,11 +87,11 @@ public class MainController {
     }
 
     private void create(){
-        int voucherSeq = MainConsole.selectCreateType();
+        int voucherSeq = Console.selectCreateType();
 
         VoucherType voucherType = voucherService.seqToType(voucherSeq);
 
-        int discountDegree = MainConsole.putDiscountDegree(voucherType);
+        int discountDegree = Console.putDiscountDegree(voucherType);
 
         voucherService.create(voucherType, discountDegree);
     }
@@ -99,7 +108,53 @@ public class MainController {
         Console.printList(blacklist);
     }
 
-    private void wallet(){
-        walletController.run();
+
+    private void runWallet(){
+        String command = readWalletCommand();
+
+        executeWalletCommand(command);
+        success(command);
+    }
+
+    private void executeWalletCommand(String command){
+        switch (command){
+            case ALLOCATE -> allocate();
+            case DELETE -> delete();
+            case SHOW_CUSTOMER_BY_VOUCHER -> showVoucherByCustomer();
+            case SHOW_VOUCHER_BY_CUSTOMER -> showCustomerByVoucher();
+            case BACK -> {}
+            default -> {
+                log.warn("invalid command. now command = {}", command);
+                throw new IllegalArgumentException("Invalid command. Type command again.");
+            }
+        }
+    }
+
+    private void allocate(){
+        UUID customerId = typeCustomerId();
+        UUID voucherId = typeVoucherId();
+
+        walletService.allocate(customerId, voucherId);
+    }
+
+    private void delete(){
+        UUID customerId = typeCustomerId();
+        UUID voucherId = typeVoucherId();
+
+        walletService.delete(customerId, voucherId);
+    }
+
+    private void showVoucherByCustomer(){
+        UUID customerId = typeCustomerId();
+        List<VoucherPolicy> vouchers = walletService.searchVouchersFromCustomer(customerId);
+
+        Console.printList(vouchers);
+    }
+
+    private void showCustomerByVoucher(){
+        UUID voucherId = typeVoucherId();
+        List<Customer> customers = walletService.searchCustomerFromVoucher(voucherId);
+
+        Console.printList(customers);
     }
 }
